See extended changes
Had to start over with test cases so I made a new file "ATMtestcases". Tested everything in lab4 pdf and a few extras in the junit. Updated quite a bit in ATM and Bank to make it testable. Consolidated the validate into 1 ATM call. More secure this way and testable Fixed a bug in account when withdraw == balance (previously wouldn't let you do that). Probably forgot some things but take a look...